vrbot-ultra-agent/requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
PyGithub>=2.1.0
lxml>=5.0.0
schedule>=1.2.0
